% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESizeCoHedg.dnsty.R
\name{ESizeCoHedg.dnsty}
\alias{ESizeCoHedg.dnsty}
\title{Calculating the Effect Size of Cohen's d & Hedges'g with the density curve}
\usage{
ESizeCoHedg.dnsty()
}
\value{
Effect Size of Cohen's d & Hedges'g
}
\description{
\code{ESizeCoHedg} calculate the Effect Size & draw the density curve
}
\examples{
# ESizeCoHedg.dnsty()
}
